Threads and Synchronization
The bookstore sells items to students, faculty/staff, and the public. There are M visitors and N cashiers in the bookstore at any given time. Visitors typically wander around shopping for items then line up to cash out. Cashiers service visitors on a first-come-first-served basis. Write a program to synchronize the visitors and cashiers using semaphores and/or monitors.